+++ src/app/index.ts
@@ -1,2 +1,1 @@
 export * from './jimpex';
-export * from './resources';
