+++ src/types/express.ts
@@ -1,0 +1,5 @@
+import type { RequestHandler, ErrorRequestHandler } from 'express';
+
+export type { Express, Router } from 'express';
+
+export type ExpressMiddleware = RequestHandler | ErrorRequestHandler;
