+++ src/types/utils.ts
@@ -1,0 +1,17 @@
+export type Dict = Record<string, unknown>;
+
+export type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends (infer U)[]
+    ? DeepPartial<U>[]
+    : T[P] extends object
+    ? DeepPartial<T[P]>
+    : T[P];
+};
+
+export type DeepReadonly<T> = {
+  readonly [P in keyof T]: T[P] extends (infer U)[]
+    ? DeepReadonly<U>[]
+    : T[P] extends object
+    ? DeepReadonly<T[P]>
+    : T[P];
+};
