+++ src/types/wootils.ts
@@ -1,0 +1,3 @@
+export type { PathUtils } from '@homer0/path-utils';
+export type { SimpleLogger } from '@homer0/simple-logger';
+export type { SimpleConfig } from '@homer0/simple-config';
